Adding reliable messaging and refactoring Client
Summary:
- Added PingMessage with resendMissing flag which allows
client to request for any messages it might not have seen
on reconnect.
- Added ability for client to resend it's non-acked message
queue when request by the server when receiving a PingMessage
with resendMissing == true for the Websocket transport adapter.
- Refactored processing and state related specifically to the
Session to live in the Session class as opposed to the Client
class which more closely mimics operation on the server.

Reviewers: tuomas

Reviewed By: tuomas

Differential Revision: https://code.uberinternal.com/D38307